+++ libs/feature/settings/src/lib/settings/settings.page.spec.ts
@@ -24,6 +24,7 @@ const SECTIONS = [
   'gifs',
   'shortcuts',
   'experimental',
+  'advanced',
 ];
 
 const ROUTES: Routes = [